backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
supabase==1.0.3
httpx>=0.23.0,<0.24.0
pytest==7.4.3
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-semantic-conventions==0.40b0
python-dotenv==1.0.0
mangum==0.17.0
redis==4.6.0 